legal-rag/requirements.txt
fastapi
uvicorn
python-multipart
celery
redis
langchain
langchain-google-genai
langchain-pinecone
langchain-community
pymupdf
pinecone-client
python-dotenv
